added binary buffers for WebGL graphics
* added JavaScript implementation of a BufferResourcesPreloader using XHR
 to retrieve binary buffers
* added bufferData and bufferSubData methods to use BufferResources
* adapted webgl example to demonstrate binary buffers
** added binary buffers, try using a url argument "?binaryBuffers"
** made OBJ reader a little more robust (comments, texture coordinates, no
exit on failure)

+++ examples/webgl/readObj.h
@@ -8,7 +8,7 @@
 // Function to read a very limited subset of obj files. Only reads
 // files with triangles. Supports v, vn, t and f.
 void readObj(const std::string &fname,
-             std::vector<double> &data);
+             std::vector<float> &data);
 
 #endif
